Use Authorization header to identity users
Expecting a cookie is not sensible because we are on a different domain
from dot com. Instead, use the standard Authorization header to pass
credentials. This is both more standard that using a query parameter (as
DAPI does), and more secure (less likely to log or otherwise lead
credentials).

The required format is:

    Authorization: Bearer [cookie/token]

An example CURL is:

    curl '../v1/avatars' \
    -X POST \
    -H 'Authorization: Bearer [cookie]' \
    -F 'image=@avatar.gif'