Migration script for moving avatars from R2 docroot to Avatar service
Script queries Discussion database for approved avatars and new social
avatars and POSTs to the migration endpoint on Avatar API which then
fetches the images and copies them to S3. The status is set in Avatar
API to approved or inactive (for social) and an unused field in the
Discussion database profile table is used to track users that have
been migrated (temporary only).