Try updating sbt and scala to fix Scala Steward
Scala Steward [requires higher minimum versions of sbt and scala than
this repository is
using](https://github.com/guardian/maintaining-scala-projects/issues/7),
and I think this is why it [recently broke on this
repository](https://github.com/guardian/scala-steward-public-repos/actions/runs/10265200458/job/28400891950).

In this commit I’ve attempted to update to newer scala and sbt versions
in the hope that it would be a straightforward update. Unfortunately
there’s a clash, as a number of dependencies now depend on
org.scala-lang.modules:scala-xml at incompatible versions with the
compiler’s dependency. (I.e. the compiler depends on
org.scala-lang.modules:scala-xml_2.12:2.2.0, but some dependencies here
depend on org.scala-lang.modules:scala-xml_2.12:1.x.)

This is the initial sbt output when you update scala and sbt:

> * org.scala-lang.modules:scala-xml_2.12:2.2.0 (early-semver) is selected over {1.3.0, 1.0.6}
>     +- org.scala-lang:scala-compiler:2.12.19              (depends on 2.2.0)
>     +- org.scalaz:scalaz-core_2.12:7.1.17                 (depends on 1.0.6)
>     +- org.scalatra:scalatra_2.12:2.6.3                   (depends on 1.0.6)
>     +- org.json4s:json4s-core_2.12:3.5.2                  (depends on 1.0.6)
>     +- net.liftweb:lift-json_2.12:3.5.0                   (depends on 1.3.0)

Unfortunately, the newest full release of scalatra_2.12 on maven central
is 2.8.4, which still depends on v1 of scala-xml. Also, that lift-json
dependency is a transitive dependency of com.gu.identity:identity-crypto
and others.