feat(ci): Setup GitHub Actions for CI/CD
Previously we relied on the CI inside of `guardian/discussion-platform`
to build and deploy Avatar API but this was unreliable and required
permanent credentials to be stored in a repository secret.

This commit sets up GitHub Actions to build and deploy the Avatar API
directly from the `guardian/avatar-api` repository. The workflow is triggered
on pushes to the `main` branch and on pull requests.